+++ app/domain/lightofday/entities/inspiration.rb
