+++ app/domain/lightofday/entities/view.rb
