+++ app/domain/lightofday/init.rb
