+++ app/domain/lightofday/repositories/for.rb
