+++ app/domain/lightofday/repositories/inspirations.rb
