+++ app/domain/lightofday/repositories/views.rb
