+++ app/domain/topics/entities/topic.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: false
+
+require 'dry-types'
+require 'dry-struct'
+require_relative '../lib/topics_calculator.rb'
+
+module LightofDay
+  module Entity
+    # entity for image topic
+    class Topic < Dry::Struct
+      include Dry.Types
+      attribute :topic_id, Strict::String
+      attribute :title, Strict::String
+      attribute :slug, Strict::String
+      attribute :starts_at, Strict::Date
+      attribute :total_photos, Strict::Integer
+      attribute :description, Strict::String
+      attribute :topic_url, Strict::String
+
+      include Mixins::TopicsCalculator
+    end
+  end
+end
