+++ app/domain/topics/init.rb
